requirements.txt
torch>=2.0
biopython>=1.80
numpy
scipy
